WIP: Rename LS to 'Microsoft Python Language Server' (#2122)
* LS symbol providers

* Different ready wait

* Progress reporting

* Options hookup

* Add diagnostic trace logging setting (hidden)

* Rename analysis engine -> language server

* Fix language server binary name

* Finish renames

* Add news

* Undo

* Baseline update

* Rename and baseline

* Fix contributions

* Baseline
