+++ .gitignore
@@ -15,7 +15,7 @@ coverage/
 .venv
 pythonFiles/experimental/ptvsd/**
 debug_coverage*/**
-analysis/**
+languageServer/**
 bin/**
 obj/**
 .pytest_cache
