+++ .vscodeignore
@@ -26,7 +26,7 @@ yarn.lock
 .nvm/**
 .vscode/**
 .vscode-test/**
-analysis/publish*.*
+languageServer/publish*.*
 bin/**
 BuildOutput/**
 coverage/**
