+++ news/1 Enhancements/2107.md
@@ -1,0 +1,2 @@
+Messages changes to reflect name of the language server: 'Microsoft Python Language Server'.
+Folder name changed from 'analysis' to 'languageServer'.
