--- src/client/activation/analysisEngineHashes.ts
@@ -1,10 +1,0 @@
-// Copyright (c) Microsoft Corporation. All rights reserved.
-// Licensed under the MIT License.
-
-// This file will be replaced by a generated one during the release build
-// with actual hashes of the uploaded packages.
-// Values are for test purposes only
-export const analysis_engine_win_x86_sha512 = 'win-x86';
-export const analysis_engine_win_x64_sha512 = 'win-x64';
-export const analysis_engine_osx_x64_sha512 = 'osx-x64';
-export const analysis_engine_linux_x64_sha512 = 'linux-x64';
