+++ src/test/analysisEngineTest.ts
@@ -6,11 +6,11 @@ import * as path from 'path';
 
 process.env.CODE_TESTS_WORKSPACE = path.join(__dirname, '..', '..', 'src', 'test');
 process.env.IS_CI_SERVER_TEST_DEBUGGER = '';
-process.env.VSC_PYTHON_ANALYSIS = '1';
+process.env.VSC_PYTHON_LANGUAGE_SERVER = '1';
 
 function start() {
     console.log('*'.repeat(100));
-    console.log('Start Analysis Engine tests');
+    console.log('Start Language Server tests');
     require('../../node_modules/vscode/bin/test');
 }
 start();
