+++ src/test/autocomplete/base.test.ts
@@ -8,7 +8,7 @@ import * as path from 'path';
 import * as vscode from 'vscode';
 import { IConfigurationService } from '../../client/common/types';
 import { rootWorkspaceUri } from '../common';
-import { closeActiveWindows, initialize, initializeTest, IsAnalysisEngineTest } from '../initialize';
+import { closeActiveWindows, initialize, initializeTest, IsLanguageServerTest } from '../initialize';
 import { UnitTestIocContainer } from '../unittests/serviceRegistry';
 
 const autoCompPath = path.join(__dirname, '..', '..', '..', 'src', 'test', 'pythonFiles', 'autocomp');
@@ -103,8 +103,8 @@ suite('Autocomplete', function () {
 
     // https://github.com/DonJayamanne/pythonVSCode/issues/630
     test('For "abc.decorators"', async () => {
-        // Disabled for MS Python Code Analysis, see https://github.com/Microsoft/PTVS/issues/3857
-        if (IsAnalysisEngineTest()) {
+        // Disabled for the Language Server, see https://github.com/Microsoft/PTVS/issues/3857
+        if (IsLanguageServerTest()) {
             return;
         }
         const textDocument = await vscode.workspace.openTextDocument(fileDecorator);
@@ -203,9 +203,9 @@ suite('Autocomplete', function () {
 
     // https://github.com/Microsoft/vscode-python/issues/110
     test('Suppress in strings/comments', async () => {
-        // Excluded from MS Python Code Analysis b/c skipping of strings and comments
+        // Excluded from the Language Server b/c skipping of strings and comments
         // is not yet there. See https://github.com/Microsoft/PTVS/issues/3798
-        if (IsAnalysisEngineTest()) {
+        if (IsLanguageServerTest()) {
             return;
         }
         const positions = [
