+++ src/test/definitions/navigation.test.ts
@@ -4,7 +4,7 @@
 import * as assert from 'assert';
 import * as path from 'path';
 import * as vscode from 'vscode';
-import { isPythonAnalysisEngineTest } from '../../client/common/constants';
+import { isLanguageServerTest } from '../../client/common/constants';
 import { closeActiveWindows, initialize, initializeTest } from '../initialize';
 
 const decoratorsPath = path.join(__dirname, '..', '..', '..', 'src', 'test', 'pythonFiles', 'definition', 'navigation');
@@ -53,82 +53,82 @@ suite('Definition Navigation', () => {
         fileDefinitions,
         new vscode.Position(2, 6),
         [fileDefinitions],
-        isPythonAnalysisEngineTest() ? [new vscode.Range(2, 4, 2, 16)] : [new vscode.Range(2, 0, 11, 17)]
+        isLanguageServerTest() ? [new vscode.Range(2, 4, 2, 16)] : [new vscode.Range(2, 0, 11, 17)]
     ));
 
     test('Nested function', buildTest(
         fileDefinitions,
         new vscode.Position(11, 16),
         [fileDefinitions],
-        isPythonAnalysisEngineTest() ? [new vscode.Range(6, 8, 6, 15)] : [new vscode.Range(6, 4, 10, 16)]
+        isLanguageServerTest() ? [new vscode.Range(6, 8, 6, 15)] : [new vscode.Range(6, 4, 10, 16)]
     ));
 
     test('Decorator usage', buildTest(
         fileDefinitions,
         new vscode.Position(13, 1),
         [fileDefinitions],
-        isPythonAnalysisEngineTest() ? [new vscode.Range(2, 4, 2, 16)] : [new vscode.Range(2, 0, 11, 17)]
+        isLanguageServerTest() ? [new vscode.Range(2, 4, 2, 16)] : [new vscode.Range(2, 0, 11, 17)]
     ));
 
     test('Function decorated by stdlib', buildTest(
         fileDefinitions,
         new vscode.Position(29, 6),
         [fileDefinitions],
-        isPythonAnalysisEngineTest() ? [new vscode.Range(21, 4, 21, 22)] : [new vscode.Range(21, 0, 27, 17)]
+        isLanguageServerTest() ? [new vscode.Range(21, 4, 21, 22)] : [new vscode.Range(21, 0, 27, 17)]
     ));
 
     test('Function decorated by local decorator', buildTest(
         fileDefinitions,
         new vscode.Position(30, 6),
         [fileDefinitions],
-        isPythonAnalysisEngineTest() ? [new vscode.Range(14, 4, 14, 9)] : [new vscode.Range(14, 0, 18, 7)]
+        isLanguageServerTest() ? [new vscode.Range(14, 4, 14, 9)] : [new vscode.Range(14, 0, 18, 7)]
     ));
 
     test('Module imported decorator usage', buildTest(
         fileUsages,
         new vscode.Position(3, 15),
         [fileDefinitions],
-        isPythonAnalysisEngineTest() ? [new vscode.Range(2, 4, 2, 16)] : [new vscode.Range(2, 0, 11, 17)]
+        isLanguageServerTest() ? [new vscode.Range(2, 4, 2, 16)] : [new vscode.Range(2, 0, 11, 17)]
     ));
 
     test('Module imported function decorated by stdlib', buildTest(
         fileUsages,
         new vscode.Position(11, 19),
         [fileDefinitions],
-        isPythonAnalysisEngineTest() ? [new vscode.Range(21, 4, 21, 22)] : [new vscode.Range(21, 0, 27, 17)]
+        isLanguageServerTest() ? [new vscode.Range(21, 4, 21, 22)] : [new vscode.Range(21, 0, 27, 17)]
     ));
 
     test('Module imported function decorated by local decorator', buildTest(
         fileUsages,
         new vscode.Position(12, 19),
         [fileDefinitions],
-        isPythonAnalysisEngineTest() ? [new vscode.Range(14, 4, 14, 9)] : [new vscode.Range(14, 0, 18, 7)]
+        isLanguageServerTest() ? [new vscode.Range(14, 4, 14, 9)] : [new vscode.Range(14, 0, 18, 7)]
     ));
 
     test('Specifically imported decorator usage', buildTest(
         fileUsages,
         new vscode.Position(7, 1),
-        isPythonAnalysisEngineTest() ? [fileUsages, fileDefinitions] : [fileDefinitions],
-        isPythonAnalysisEngineTest()
-            ? [new vscode.Range(1, 45, 1, 57), new vscode.Range(2, 4, 2, 16)]
+        isLanguageServerTest() ? [fileDefinitions] : [fileDefinitions],
+        isLanguageServerTest()
+            ? [new vscode.Range(2, 4, 2, 16)]
             : [new vscode.Range(2, 0, 11, 17)]
     ));
 
     test('Specifically imported function decorated by stdlib', buildTest(
         fileUsages,
         new vscode.Position(14, 6),
-        isPythonAnalysisEngineTest() ? [fileUsages, fileDefinitions] : [fileDefinitions],
-        isPythonAnalysisEngineTest()
-            ? [new vscode.Range(1, 25, 1, 43), new vscode.Range(21, 4, 21, 22)]
+        isLanguageServerTest() ? [fileDefinitions] : [fileDefinitions],
+        isLanguageServerTest()
+            ? [new vscode.Range(21, 4, 21, 22)]
             : [new vscode.Range(21, 0, 27, 17)]
     ));
 
     test('Specifically imported function decorated by local decorator', buildTest(
         fileUsages,
         new vscode.Position(15, 6),
-        isPythonAnalysisEngineTest() ? [fileUsages, fileDefinitions] : [fileDefinitions],
-        isPythonAnalysisEngineTest()
-            ? [new vscode.Range(1, 59, 1, 64), new vscode.Range(14, 4, 14, 9)]
+        isLanguageServerTest() ? [fileDefinitions] : [fileDefinitions],
+        isLanguageServerTest()
+            ? [new vscode.Range(14, 4, 14, 9)]
             : [new vscode.Range(14, 0, 18, 7)]
     ));
 });
